Basic Monte-Carlo rendering,

+++ RayTracing/vec3.h
@@ -64,7 +64,7 @@ inline std::ostream& operator<<(std::ostream &os, const vec3 &t) {
 }
 
 inline void vec3::make_unit_vector() {
-	float k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
+	double k = 1.0 / sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
 	e[0] *= k; e[1] *= k; e[2] *= k;
 }
 
@@ -102,7 +102,7 @@ inline vec3 operator/(vec3 v, float t) {
 }
 
 //Dot product
-inline float dot(const vec3 &v1, const vec3 &v2) {
+inline double dot(const vec3 &v1, const vec3 &v2) {
 	return v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2];
 }
 
@@ -149,7 +149,7 @@ inline vec3& vec3::operator*=(float t) {
 }
 
 inline vec3& vec3::operator/=(float t) {
-	float k = 1.0 / t;
+	double k = 1.0 / t;
 
 	e[0] *= k;
 	e[1] *= k;
